OPENNLP-912: Move rules into rules.xml
