+++ opennlp-tools/src/main/resources/opennlp/tools/sentdetect/segment/rules.xml
@@ -1,0 +1,131 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!--
+	Licensed to the Apache Software Foundation (ASF) under one
+	or more contributor license agreements.  See the NOTICE file
+	distributed with this work for additional information
+	regarding copyright ownership.  The ASF licenses this file
+	to you under the Apache License, Version 2.0 (the
+	"License"); you may not use this file except in compliance
+	with the License.  You may obtain a copy of the License at
+
+	http://www.apache.org/licenses/LICENSE-2.0
+
+	Unless required by applicable law or agreed to in writing,
+	software distributed under the License is distributed on an
+	"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+	KIND, either express or implied.  See the License for the
+	specific language governing permissions and limitations
+	under the License.
+-->
+
+<languageRules>
+  <languageRule name="eng">
+
+    <!--common-->
+    <rule break="yes">
+      <beforeBreak>\n</beforeBreak>
+      <afterBreak></afterBreak>
+    </rule>
+    <rule break="yes">
+      <beforeBreak> </beforeBreak>
+      <afterBreak>\n</afterBreak>
+    </rule>
+    <rule break="yes">
+      <beforeBreak>[\.\?!]+\s+</beforeBreak>
+      <afterBreak>[^\.]</afterBreak>
+    </rule>
+    <rule break="yes">
+      <beforeBreak>[\.\?!]+</beforeBreak>
+      <afterBreak>\s*(A |Being|Did|For|He|How|However|I|In|It|Millions|More|She|That|The|There|They|We|What|When|Where|Who|Why)</afterBreak>
+    </rule>
+    <rule break="yes">
+      <beforeBreak>[!?\.-][\"\'“”]\s+</beforeBreak>
+      <afterBreak>[A-Z]</afterBreak>
+    </rule>
+    <rule break="yes">
+      <beforeBreak>(?&lt;=\S)(!|\?){3,}</beforeBreak>
+      <afterBreak>(?=(\s|\Z|$))</afterBreak>
+    </rule>
+    <rule break="no">
+      <beforeBreak>[\.\?!]+\s*</beforeBreak>
+      <afterBreak>(?=[\.\?!])</afterBreak>
+    </rule>
+    <rule break="no">
+      <beforeBreak>([a-zA-z]°)\.\s*</beforeBreak>
+      <afterBreak>(?=\d+)</afterBreak>
+    </rule>
+    <rule break="no">
+      <beforeBreak>\s</beforeBreak>
+      <afterBreak>(?=[a-z])</afterBreak>
+    </rule>
+
+    <!-- number -->
+    <rule break="no">
+      <beforeBreak>\d\.</beforeBreak>
+      <afterBreak>(?=\d)</afterBreak>
+    </rule>
+
+    <!-- name -->
+    <rule break="no">
+      <beforeBreak>(Mr|Mrs|Ms|Dr|p.m|a.m|tel)\.</beforeBreak>
+      <afterBreak>\s*</afterBreak>
+    </rule>
+    <rule break="yes">
+      <beforeBreak>(P\.M\.|A\.M\.)</beforeBreak>
+      <afterBreak>\s+</afterBreak>
+    </rule>
+    <rule break="no">
+      <beforeBreak>(?&lt;=(?&lt;=^)[A-Z]\.\s+|(?&lt;=\A)[A-Z]\.\s+|[A-Z]\.\s+|(?&lt;=^)[A-Z][a-z]\.\s+|(?&lt;=\A)[A-Z][a-z]\.\s+|(?&lt;=\s)[A-Z][a-z]\.\s)</beforeBreak>
+      <afterBreak>(?!(A |Being|Did|For|He|How|However|I|In|It|Millions|More|She|That|The|There|They|We|What|When|Where|Who|Why))</afterBreak>
+    </rule>
+
+    <!-- betweenPunctuation -->
+    <rule break="no">
+      <beforeBreak>(?&lt;=\s)'(?:[^']|'[a-zA-Z])*'</beforeBreak>
+      <afterBreak></afterBreak>
+    </rule>
+    <rule break="no">
+      <beforeBreak>(?&lt;=\s)‘(?:[^’]|’[a-zA-Z])*’</beforeBreak>
+      <afterBreak></afterBreak>
+    </rule>
+    <rule break="no">
+      <beforeBreak>"(?>[^"\\]+|\\{2}|\\.)*"</beforeBreak>
+      <afterBreak></afterBreak>
+    </rule>
+    <rule break="no">
+      <beforeBreak>«(?>[^»\\]+|\\{2}|\\.)*»</beforeBreak>
+      <afterBreak></afterBreak>
+    </rule>
+    <rule break="no">
+      <beforeBreak>“(?>[^”\\]+|\\{2}|\\.)*”</beforeBreak>
+      <afterBreak></afterBreak>
+    </rule>
+    <rule break="no">
+      <beforeBreak>\[(?>[^\]\\]+|\\{2}|\\.)*\]</beforeBreak>
+      <afterBreak></afterBreak>
+    </rule>
+    <rule break="no">
+      <beforeBreak>\((?>[^\(\)\\]+|\\{2}|\\.)*\)</beforeBreak>
+      <afterBreak></afterBreak>
+    </rule>
+    <rule break="no">
+      <beforeBreak>(?&lt;=\s)\-\-(?>[^\-\-])*\-\-</beforeBreak>
+      <afterBreak></afterBreak>
+    </rule>
+
+    <!-- list -->
+    <rule break="no">
+      <beforeBreak>((?&lt;=^)[a-z]\.|(?&lt;=\A)[a-z]\.|(?&lt;=\s)[a-z]\.)</beforeBreak>
+      <afterBreak>\s*(?!(A |Being|Did|For|He|How|However|I|In|It|Millions|More|She|That|The|There|They|We|What|When|Where|Who|Why))</afterBreak>
+    </rule>
+    <rule break="no">
+      <beforeBreak>(?&lt;=\s)\d{1,2}\.(\s)|^\d{1,2}\.(\s)|(?&lt;=\s)\d{1,2}\.(\))|^\d{1,2}\.(\))|(?&lt;=\s\-)\d{1,2}\.(\s)|(?&lt;=^\-)\d{1,2}\.(\s)|(?&lt;=\s\⁃)\d{1,2}\.(\s)|(?&lt;=^\⁃)\d{1,2}\.(\s)|(?&lt;=\s\-)\d{1,2}\.(\))|(?&lt;=^\-)\d{1,2}\.(\))|(?&lt;=\s\⁃)\d{1,2}\.(\))|(?&lt;=^\⁃)\d{1,2}\.(\))|(\•)\s*\d{1,2}\.(\s)|(?&lt;=\s)\d{1,2}(\))</beforeBreak>
+      <afterBreak>\s*</afterBreak>
+    </rule>
+    <rule break="yes">
+      <beforeBreak></beforeBreak>
+      <afterBreak>\s+((?&lt;=\s)\d{1,2}\.(?=\s)|^\d{1,2}\.(?=\s)|(?&lt;=\s)\d{1,2}\.(?=\))|^\d{1,2}\.(?=\))|((?&lt;=\s)\-)\d{1,2}\.(?=\s)|(^\-)\d{1,2}\.(?=\s)|((?&lt;=\s)\⁃)\d{1,2}\.(?=\s)|(^\⁃)\d{1,2}\.(?=\s)|((?&lt;=\s)\-)\d{1,2}\.(?=\))|(^\-)\d{1,2}\.(?=\))|((?&lt;=\s)\⁃)\d{1,2}\.(?=\))|(^\⁃)\d{1,2}\.(?=\))|(\•)\s*\d{1,2}\.(\s)|(?&lt;=\s)\d{1,2}(?=\)))</afterBreak>
+    </rule>
+  </languageRule>
+</languageRules>
